+++ zapit/+zapit/+settings/returnZapitSettingsFileName.m
@@ -2,7 +2,7 @@
     % Return the zapit settings file name as a string
     %
     % Purpose
-    % Multiple files need access to this file name so we difine it here once.
+    % Multiple files need access to this file name so we define it here once.
     % Returns "zapitSystemSettings.yml"
     %
     % Rob Campbell - SWC 2022