+++ zapit/+zapit/@stimConfig/filterForEphys.m
@@ -1,5 +1,5 @@
-function ephysWaveform = filterForEphys(obj,inputWaveform)
-    % Filter the laser waveform for ephys
+function [ephysWaveform,scaleFactor] = filterForEphys(obj,inputWaveform)
+    % Filter the laser waveform for ephys: turning square waves into sinusoids
     %
     % function ephysWaveform = zapit.stimConfig.filterForEphys(obj,inputWaveform)
     %
@@ -10,36 +10,26 @@ function ephysWaveform = filterForEphys(obj,inputWaveform)
     % overall amplitude of the signal is increased such that the area under the curve
     % remains the same. Thus the total number of photons delivered to the sample remain
     % unaltered.
-    % The ephys waveform is made by filtering the raw laser signal with a Hanning filter
-    % and re-zeroing all values that were zero originally.
+    % The ephys waveform is made by converting the "on" epochs into a half cycle of
+    % sinusoid.
     %
+    % NOTE
+    % The waveform will have a higher peak value (>1) and this should take into account
+    % the higher laser power that we need to produce. See how zapit.pointer.sendSamples
+    % produces the laser power.
     %
     % Inputs
     % inputWaveform - The raw laser waveform before filtering.
     %
     % Outputs
     % ephysWaveform - The filtered waveform.
+    % scaleFactor [optional] - Scalar indicating by how much the waveform was multiplied such that
+    %       it has the same area as the original trace. So takes into account the lower
+    %       integrated power of the sinusoidal waveform.
+    %
     %
     % Rob Campbell - SWC 2023
 
-
-
-    if isempty(obj.parent)
-        return
-    end
-
-
-    % To filter we need to ensure the *on* periods are centred. Therefore we circshift by
-    % half the blanking time if the first sample is not a zero
-    if inputWaveform(1)>0
-        blankingTime = obj.parent.settings.experiment.blankingTime_ms;
-        samplesPerSecond = obj.parent.DAQ.samplesPerSecond;
-        cShiftQuantity = samplesPerSecond * blankingTime*1E-3 * 0.5;
-
-        inputWaveform = circshift(inputWaveform, -cShiftQuantity);
-    else
-        cShiftQuantity = 0;
-    end
 
 
     % Find the length of the on and off blocks. We know the stimulus "on" times will be
@@ -47,48 +37,42 @@ function ephysWaveform = filterForEphys(obj,inputWaveform)
     blockLengths = diff(find(abs(diff(inputWaveform))));
     maxBlockLength = max(blockLengths);
 
+    % The start of the stim epoch blocks
+    f=find(diff(inputWaveform)==1);
 
-    % Define the tapering filter (e.g., Hann window) based on the length of the "on" block.
-    taperLength = round(maxBlockLength/4);
-    taper = hann(taperLength * 2)';
+    % Waveform
+    wForm = sin(linspace(0,pi,maxBlockLength));
+    maxVal = 1;
+    wForm(wForm>maxVal)=maxVal;
 
-    % Create the modified signal
-    ephysWaveform = conv(inputWaveform, taper, 'same');
-
-    % Find the indices where zeros should be placed
-    zeroIndices = inputWaveform == 0;
-
-    % Replace the values in the output signal where zeros should be present
-    ephysWaveform(zeroIndices) = 0;
-
-
-    % Shift back
-
-    ephysWaveform = circshift(ephysWaveform, cShiftQuantity);
-
+    % Populate
+    ephysWaveform = zeros(size(inputWaveform));
+    for ii=1:length(f)
+        s = f(ii)+1;
+        ephysWaveform(s:s+maxBlockLength-1) = wForm;
+    end
 
 
 
 
     % scale such that that area under the curves are the same
-    E=(sum(ephysWaveform));
-    I=(sum(inputWaveform));
+    E = sum(ephysWaveform);
+    I = sum(inputWaveform);
+    scaleFactor = I/E;
 
-    ephysWaveform = ephysWaveform / (E/I);
+    % Check that the scaling is correct (it must be, but let's be paranoid)
+    ephysWaveform = ephysWaveform * scaleFactor;
 
     if round(sum(ephysWaveform)) ~= round(sum(inputWaveform))
-        fprintf('Warning! The areas under the original and ephys-corrected waveforms are not the same!\n')
+        fprintf('zapit.stimConfig.%s -- Warning! The areas under the original and ephys-corrected waveforms are not the same!\n', mfilename)
     end
 
-    doPlot = false;
+    doPlot = true;
     if doPlot
-        inputWaveform = circshift(inputWaveform, cShiftQuantity);
         plot(ephysWaveform,'-k')
         hold on
         plot(inputWaveform,'-r')
         hold off
     end
-
-
 
 end %filterForEphys